Add Concern to Fix Metrics/ClassLength
