+++ .rubocop_todo.yml
@@ -525,7 +525,6 @@ Metrics/ClassLength:
     - 'app/controllers/admin/enterprise_fees_controller.rb'
     - 'app/controllers/admin/enterprises_controller.rb'
     - 'app/controllers/admin/order_cycles_controller.rb'
-    - 'app/controllers/admin/product_import_controller.rb'
     - 'app/controllers/admin/resource_controller.rb'
     - 'app/controllers/admin/schedules_controller.rb'
     - 'app/controllers/admin/subscriptions_controller.rb'
